refactor: experiment_logic.py, organize functions
